VSH_Project_MVP/requirements.txt
fastmcp
anthropic
semgrep
fastapi
uvicorn
tree-sitter
tree-sitter-python
python-dotenv
